Fix rebase glitches
