+++ sdks/java/io/amazon-web-services2/src/test/java/org/apache/beam/sdk/io/aws2/kinesis/KinesisIOOptionsTest.java
@@ -17,7 +17,7 @@
  */
 package org.apache.beam.sdk.io.aws2.kinesis;
 
-import static org.apache.beam.sdk.io.aws2.kinesis.Helpers.createIOOptions;
+import static org.apache.beam.sdk.io.aws2.kinesis.TestHelpers.createIOOptions;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Map;
