+++ sdks/java/io/amazon-web-services2/src/test/java/org/apache/beam/sdk/io/aws2/kinesis/ShardReadersPoolExtendedTest.java
@@ -17,11 +17,11 @@
  */
 package org.apache.beam.sdk.io.aws2.kinesis;
 
-import static org.apache.beam.sdk.io.aws2.kinesis.Helpers.mockRecords;
-import static org.apache.beam.sdk.io.aws2.kinesis.Helpers.mockShardIterators;
-import static org.apache.beam.sdk.io.aws2.kinesis.Helpers.recordWithMinutesAgo;
 import static org.apache.beam.sdk.io.aws2.kinesis.TestHelpers.createAggregatedRecords;
 import static org.apache.beam.sdk.io.aws2.kinesis.TestHelpers.createRecords;
+import static org.apache.beam.sdk.io.aws2.kinesis.TestHelpers.mockRecords;
+import static org.apache.beam.sdk.io.aws2.kinesis.TestHelpers.mockShardIterators;
+import static org.apache.beam.sdk.io.aws2.kinesis.TestHelpers.recordWithMinutesAgo;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
